backend/requirements.txt
# backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
redis==5.0.1
rq==1.15.1
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
umap-learn==0.5.4
python-dotenv==1.0.0
